+++ meta/regexp.rb
@@ -4,44 +4,39 @@ Mutant::Meta::Example.add :regexp do
   source '/foo/'
 
   singleton_mutations
-
-  mutation '//'
-  mutation '/nomatch\A/'
+  regexp_mutations
 end
 
 Mutant::Meta::Example.add :regexp do
   source '/#{foo.bar}n/'
 
   singleton_mutations
+  regexp_mutations
 
   mutation '/#{foo}n/'
-  mutation '/#{nil}n/'
   mutation '/#{self.bar}n/'
+  mutation '/#{nil}n/'
   mutation '/#{self}n/'
-  mutation '//'
-  mutation '/nomatch\A/'
 end
 
 Mutant::Meta::Example.add :regexp do
   source '/#{foo}/'
 
   singleton_mutations
+  regexp_mutations
 
   mutation '/#{self}/'
   mutation '/#{nil}/'
-  mutation '//'
-  mutation '/nomatch\A/'
 end
 
 Mutant::Meta::Example.add :regexp do
   source '/#{foo}#{nil}/'
 
   singleton_mutations
+  regexp_mutations
 
   mutation '/#{nil}#{nil}/'
   mutation '/#{self}#{nil}/'
-  mutation '//'
-  mutation '/nomatch\A/'
 end
 
 Mutant::Meta::Example.add :regexp do
@@ -70,11 +65,37 @@ Mutant::Meta::Example.add :regexp do
   mutation 'true if /nomatch\A/'
 end
 
-# Case where MRI would accept an expression but regexp_parser not.
 Mutant::Meta::Example.add :regexp do
-  source '/u{/'
+  source '/(?(1)(foo)(bar))/'
+
+  singleton_mutations
+  regexp_mutations
+
+  mutation '/(?(1)(?:foo)(bar))/'
+  mutation '/(?(1)(foo)(?:bar))/'
+end
+
+# MRI accepts this regex but `regexp_parser` does not.
+# See: https://github.com/ammar/regexp_parser/issues/75
+Mutant::Meta::Example.add :regexp do
+  source '/\xA/'
 
   singleton_mutations
   mutation '//'
   mutation '/nomatch\A/'
 end
+
+# MRI accepts this regex but `regexp_parser` does not.
+# See: https://github.com/ammar/regexp_parser/issues/76
+Mutant::Meta::Example.add :regexp do
+  source '/(?<æ>.)/'
+
+  singleton_mutations
+  mutation '//'
+  mutation '/nomatch\A/'
+end
+
+Pathname
+  .glob(Pathname.new(__dir__).join('regexp', '*.rb'))
+  .sort
+  .each(&Kernel.public_method(:require))